Refactor window and input code
- KeyboardInput and PointerInput are owned by Window because they only
get input/events from a single window
- KeyboardInput uses event processing instead of polling to read key
state

+++ src/Camera.hpp
@@ -41,7 +41,7 @@ struct Camera
 		return Mat4x4f(transform.rotation.GetTransposed()) * Matrix::Translate(-(transform.position));
 	}
 
-	inline void SetAspectRatio(int width, int height)
+	inline void SetAspectRatio(float width, float height)
 	{
 		aspectRatio = width / static_cast<float>(height);
 	}
